moving to src folder, addressing nullify coalesce feedback, addressing import feedback
